Test/beefy light client merge upstream

+++ relayer/provider/substrate/key_test.go
@@ -2,6 +2,8 @@ package substrate_test
 
 import (
 	"testing"
+
+	"github.com/stretchr/testify/require"
 )
 
 // TestKeyRestore restores a test mnemonic
@@ -11,12 +13,37 @@ func TestKeyRestore(t *testing.T) {
 	expectedAddress := "5Hn67YZ75F3XrHiJAtiscJMGQ4zFNw9e45CNfLuxL6vEVYz8"
 
 	testProvider, err := getTestProvider()
-	_ = testProvider.DeleteKey(keyName) // Delete if test is being run again
-	address, err := testProvider.RestoreKey(keyName, mnemonic)
-	if err != nil {
-		t.Fatalf("Error while restoring mnemonic: %v", err)
+	require.Nil(t, err)
+
+	if testProvider.KeyExists(keyName) {
+		err = testProvider.DeleteKey(keyName) // Delete if test is being run again
+		require.Nil(t, err)
 	}
-	if address != expectedAddress {
-		t.Fatalf("Restored address: %s does not match expected: %s", address, expectedAddress)
+
+	address, err := testProvider.RestoreKey(keyName, mnemonic, 0)
+	require.Nil(t, err)
+
+	require.Equal(t, expectedAddress, address, "Restored address: %s does not match expected: %s", address, expectedAddress)
+}
+
+func TestKeyInsert(t *testing.T) {
+	keyName := "test_key"
+	mnemonic := "blind master acoustic speak victory lend kiss grab glad help demand hood roast zone lend sponsor level cheap truck kingdom apology token hover reunion"
+	expectedAddress := "5Hn67YZ75F3XrHiJAtiscJMGQ4zFNw9e45CNfLuxL6vEVYz8"
+
+	testProvider, err := getTestProvider()
+	require.Nil(t, err)
+
+	if testProvider.KeyExists(keyName) {
+		err = testProvider.DeleteKey(keyName) // Delete if test is being run again
+		require.Nil(t, err)
 	}
+
+	address, err := testProvider.RestoreKey(keyName, mnemonic, 0)
+	require.Nil(t, err)
+
+	retrievedAddress, err := testProvider.ShowAddress(keyName)
+	require.Nil(t, err)
+
+	require.Equal(t, expectedAddress, address, "Restored address: %s does not match expected: %s", retrievedAddress, expectedAddress)
 }
